create context and split components
