+++ src/components/Navbar.js
@@ -1,0 +1,34 @@
+import { useState } from 'react'
+import { NavLink } from 'react-router-dom'
+
+import { GiHamburgerMenu } from 'react-icons/gi'
+import { AiOutlineClose } from 'react-icons/ai'  
+
+export default function Navbar() {
+    const [navbarOpen, setNavbarOpen] = useState(false)
+
+    const handleNavbarToggle = () => {
+      setNavbarOpen(previous => !previous)
+    }
+
+    const closeMenu = () => {
+      setNavbarOpen(false)
+    }
+
+    return (
+      <header >
+        <NavLink to="/" className={'logo'}>
+          <img className='logo__leaves' src={require('../assets/one-planet-logo.png')} alt='One Planet logo' />
+        </NavLink>
+        <nav>
+          <button onClick={handleNavbarToggle}>{navbarOpen ? <AiOutlineClose /> : <GiHamburgerMenu />}</button>
+          <div className={`menuNav ${navbarOpen ? "showMenu" : ""}`}>
+            <NavLink onClick={() => closeMenu()} to="/blog"  >Blog</NavLink>
+            <NavLink onClick={() => closeMenu()} to="/shop" >Shop</NavLink>
+            <NavLink onClick={() => closeMenu()} to="/reviews" >Reviews</NavLink>
+            <NavLink onClick={() => closeMenu()} to="/contact" >Contact</NavLink>
+          </div>
+        </nav>
+      </header>
+    )
+  }
