+++ src/components/Reviews.js
@@ -1,0 +1,11 @@
+import { useContext } from 'react'
+import { documentToReactComponents } from '@contentful/rich-text-react-renderer'
+import { ShopContext } from '../contexts/shopContext'
+
+export default function Reviews() {
+    const { reviews } = useContext(ShopContext)
+
+    return (
+
+    )
+}
