+++ src/index.js
@@ -1,14 +1,17 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './styles/index.css';
-import App from './components/App';
-import reportWebVitals from './reportWebVitals';
+import './styles/index.css'
+import React from 'react'
+import ReactDOM from 'react-dom'
+import ShopContextProvider from './contexts/shopContext'
 import { BrowserRouter } from 'react-router-dom'
+import App from './components/App'
+import reportWebVitals from './reportWebVitals'
 
 ReactDOM.render(
-  <BrowserRouter>
-    <App />
-  </BrowserRouter>,
+  <ShopContextProvider>
+    <BrowserRouter>
+      <App />
+    </BrowserRouter>
+  </ShopContextProvider>,
   document.getElementById('root')
 );
